requirements.txt
requests==2.31.0
Pillow==10.2.0
tk==0.1.0 